requirements.txt
python-nmap
subconscious
